scrambleclip2/requirements.txt
streamlit>=1.22.0
moviepy==1.0.3
numpy>=1.22.0
Pillow>=9.0.0
opencv-python-headless>=4.6.0
watchdog>=2.1.9
scipy>=1.8.0
requests>=2.27.1
decorator>=4.0.2
tqdm>=4.11.2
imageio>=2.5.0
imageio-ffmpeg>=0.4.0
proglog>=0.1.9
psutil>=5.9.0
Wand>=0.6.13 